Add OrderStat
