+++ src/components/feed/OrderFeed.tsx
@@ -53,57 +53,16 @@ const testOrders: Array<TOrder> = [
 	},
 ];
 
-const readyOrders = [34533, 34532, 34530, 34527, 34525];
-const inProgressOrders = [34538, 34541, 34542];
-const total = 28752;
-const totalToday = 138;
-
 const OrderFeed: React.FC = () => {
 	return (
-		<section className={styles.feedWrapper}>
-			<div className={styles.ordersSection}>
-				<h1 className='text text_type_main-large mb-5'>Лента заказов</h1>
-				<div className={styles.ordersList + ' custom-scroll'}>
-					{testOrders.map((order) => (
-						<OrderCard key={order._id} order={order} />
-					))}
-				</div>
+		<div className={styles.ordersSection}>
+			<h1 className='text text_type_main-large mb-5'>Лента заказов</h1>
+			<div className={styles.ordersList + ' custom-scroll'}>
+				{testOrders.map((order) => (
+					<OrderCard key={order._id} order={order} />
+				))}
 			</div>
-			<div className={styles.statsSection}>
-				<div className={styles.statusesRow}>
-					<div>
-						<p className='text text_type_main-medium mb-2'>Готовы:</p>
-						<ul className={styles.statusList + ' ' + styles.ready}>
-							{readyOrders.map((num) => (
-								<li key={num} className='text text_type_digits-default'>
-									{num}
-								</li>
-							))}
-						</ul>
-					</div>
-					<div>
-						<p className='text text_type_main-medium mb-2'>В работе:</p>
-						<ul className={styles.statusList}>
-							{inProgressOrders.map((num) => (
-								<li key={num} className='text text_type_digits-default'>
-									{num}
-								</li>
-							))}
-						</ul>
-					</div>
-				</div>
-				<div className='mt-15'>
-					<p className='text text_type_main-medium'>Выполнено за все время:</p>
-					<span className='text text_type_digits-large'>
-						{total.toLocaleString('ru-RU')}
-					</span>
-				</div>
-				<div className='mt-15'>
-					<p className='text text_type_main-medium'>Выполнено за сегодня:</p>
-					<span className='text text_type_digits-large'>{totalToday}</span>
-				</div>
-			</div>
-		</section>
+		</div>
 	);
 };
