+++ src/components/feed/index.tsx
@@ -1,0 +1,5 @@
+import OrderCard from './OrderCard';
+import OrderFeed from './OrderFeed';
+import OrderStat from './OrderStat';
+
+export { OrderCard, OrderFeed, OrderStat };
